feat(config): add `RequiredOrThrows` exception. It can throw an exception when the target profile doesn't have a specified value.
There are also a few other changes:
The exception that needs to be caught is switched from `IOException` to `ParsedFailedException`. Change the `try-catch` statement you use to handle exceptions.
"." It is now considered a valid path. But it doesn't create a block for it. Instead, it is passed directly in the parser.